// packages/indexer/ingest-ts/src/scip/generated/scip.ts
export enum ProtocolVersion {
  UnspecifiedProtocolVersion = 0,
  UNRECOGNIZED = -1,
}
// ...
export enum TextEncoding {
  UnspecifiedTextEncoding = 0,
  UTF8 = 1,
  UTF16 = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum SymbolRole {
  /**
   * UnspecifiedSymbolRole - This case is not meant to be used; it only exists to avoid an error
   * from the Protobuf code generator.
   */
  UnspecifiedSymbolRole = 0,
  /** Definition - Is the symbol defined here? If not, then this is a symbol reference. */
  Definition = 1,
  /** Import - Is the symbol imported here? */
  Import = 2,
  /** WriteAccess - Is the symbol written here? */
  WriteAccess = 4,
  /** ReadAccess - Is the symbol read here? */
  ReadAccess = 8,
  /** Generated - Is the symbol in generated code? */
  Generated = 16,
  /** Test - Is the symbol in test code? */
  Test = 32,
  /**
   * ForwardDefinition - Is this a signature for a symbol that is defined elsewhere?
   *
   * Applies to forward declarations for languages like C, C++
   * and Objective-C, as well as `val` declarations in interface
   * files in languages like SML and OCaml.
   */
  ForwardDefinition = 64,
  UNRECOGNIZED = -1,
}
// ...
export interface Index {
  /** Metadata about this index. */
  metadata?:
    | Metadata
    | undefined;
  /** Documents that belong to this index. */
  documents?:
    | Document[]
    | undefined;
  /**
   * (optional) Symbols that are referenced from this index but are defined in
   * an external package (a separate `Index` message). Leave this field empty
   * if you assume the external package will get indexed separately. If the
   * external package won't get indexed for some reason then you can use this
   * field to provide hover documentation for those external symbols.
   */
  externalSymbols?: SymbolInformation[] | undefined;
}

export interface Metadata {
  /** Which version of this protocol was used to generate this index? */
  version?:
    | ProtocolVersion
    | undefined;
  /** Information about the tool that produced this index. */
  toolInfo?:
    | ToolInfo
    | undefined;
  /**
   * URI-encoded absolute path to the root directory of this index. All
   * documents in this index must appear in a subdirectory of this root
   * directory.
   */
  projectRoot?:
    | string
    | undefined;
  /**
   * Text encoding of the source files on disk that are referenced from
   * `Document.relative_path`. This value is unrelated to the `Document.text`
   * field, which is a Protobuf string and hence must be UTF-8 encoded.
   */
  textDocumentEncoding?: TextEncoding | undefined;
}

export interface ToolInfo {
  /** Name of the indexer that produced this index. */
  name?:
    | string
    | undefined;
  /** Version of the indexer that produced this index. */
  version?:
    | string
    | undefined;
  /** Command-line arguments that were used to invoke this indexer. */
  arguments?: string[] | undefined;
}
// ...
export interface Package {
  manager?: string | undefined;
  name?: string | undefined;
  version?: string | undefined;
}
// ...
export enum SymbolInformation_Kind {
  UnspecifiedKind = 0,
  /** AbstractMethod - A method which may or may not have a body. For Java, Kotlin etc. */
  AbstractMethod = 66,
  /** Accessor - For Ruby's attr_accessor */
  Accessor = 72,
  Array = 1,
  /** Assertion - For Alloy */
  Assertion = 2,
  AssociatedType = 3,
  /** Attribute - For C++ */
  Attribute = 4,
  /** Axiom - For Lean */
  Axiom = 5,
  Boolean = 6,
  Class = 7,
  /** Concept - For C++ */
  Concept = 86,
  Constant = 8,
  Constructor = 9,
  /** Contract - For Solidity */
  Contract = 62,
  /** DataFamily - For Haskell */
  DataFamily = 10,
  /** Delegate - For C# and F# */
  Delegate = 73,
  Enum = 11,
  EnumMember = 12,
  Error = 63,
  Event = 13,
  /** Extension - For Dart */
  Extension = 84,
  /** Fact - For Alloy */
  Fact = 14,
  Field = 15,
  File = 16,
  Function = 17,
  /** Getter - For 'get' in Swift, 'attr_reader' in Ruby */
  Getter = 18,
  /** Grammar - For Raku */
  Grammar = 19,
  /** Instance - For Purescript and Lean */
  Instance = 20,
  Interface = 21,
  Key = 22,
  /** Lang - For Racket */
  Lang = 23,
  /** Lemma - For Lean */
  Lemma = 24,
  /** Library - For solidity */
  Library = 64,
  Macro = 25,
  Method = 26,
  /** MethodAlias - For Ruby */
  MethodAlias = 74,
  /**
   * MethodReceiver - Analogous to 'ThisParameter' and 'SelfParameter', but for languages
   * like Go where the receiver doesn't have a conventional name.
   */
  MethodReceiver = 27,
  /** MethodSpecification - Analogous to 'AbstractMethod', for Go. */
  MethodSpecification = 67,
  /** Message - For Protobuf */
  Message = 28,
  /** Mixin - For Dart */
  Mixin = 85,
  /** Modifier - For Solidity */
  Modifier = 65,
  Module = 29,
  Namespace = 30,
  Null = 31,
  Number = 32,
  Object = 33,
  Operator = 34,
  Package = 35,
  PackageObject = 36,
  Parameter = 37,
  ParameterLabel = 38,
  /** Pattern - For Haskell's PatternSynonyms */
  Pattern = 39,
  /** Predicate - For Alloy */
  Predicate = 40,
  Property = 41,
  /** Protocol - Analogous to 'Trait' and 'TypeClass', for Swift and Objective-C */
  Protocol = 42,
  /** ProtocolMethod - Analogous to 'AbstractMethod', for Swift and Objective-C. */
  ProtocolMethod = 68,
  /** PureVirtualMethod - Analogous to 'AbstractMethod', for C++. */
  PureVirtualMethod = 69,
  /** Quasiquoter - For Haskell */
  Quasiquoter = 43,
  /** SelfParameter - 'self' in Python, Rust, Swift etc. */
  SelfParameter = 44,
  /** Setter - For 'set' in Swift, 'attr_writer' in Ruby */
  Setter = 45,
  /** Signature - For Alloy, analogous to 'Struct'. */
  Signature = 46,
  /** SingletonClass - For Ruby */
  SingletonClass = 75,
  /** SingletonMethod - Analogous to 'StaticMethod', for Ruby. */
  SingletonMethod = 76,
  /** StaticDataMember - Analogous to 'StaticField', for C++ */
  StaticDataMember = 77,
  /** StaticEvent - For C# */
  StaticEvent = 78,
  /** StaticField - For C# */
  StaticField = 79,
  /** StaticMethod - For Java, C#, C++ etc. */
  StaticMethod = 80,
  /** StaticProperty - For C#, TypeScript etc. */
  StaticProperty = 81,
  /** StaticVariable - For C, C++ */
  StaticVariable = 82,
  String = 48,
  Struct = 49,
  /** Subscript - For Swift */
  Subscript = 47,
  /** Tactic - For Lean */
  Tactic = 50,
  /** Theorem - For Lean */
  Theorem = 51,
  /**
   * ThisParameter - Method receiver for languages
   * 'this' in JavaScript, C++, Java etc.
   */
  ThisParameter = 52,
  /** Trait - Analogous to 'Protocol' and 'TypeClass', for Rust, Scala etc. */
  Trait = 53,
  /** TraitMethod - Analogous to 'AbstractMethod', for Rust, Scala etc. */
  TraitMethod = 70,
  /**
   * Type - Data type definition for languages like OCaml which use `type`
   * rather than separate keywords like `struct` and `enum`.
   */
  Type = 54,
  TypeAlias = 55,
  /** TypeClass - Analogous to 'Trait' and 'Protocol', for Haskell, Purescript etc. */
  TypeClass = 56,
  /** TypeClassMethod - Analogous to 'AbstractMethod', for Haskell, Purescript etc. */
  TypeClassMethod = 71,
  /** TypeFamily - For Haskell */
  TypeFamily = 57,
  TypeParameter = 58,
  /** Union - For C, C++, Capn Proto */
  Union = 59,
  Value = 60,
  /**
   * Variable - Next = 87;
   * Feel free to open a PR proposing new language-specific kinds.
   */
  Variable = 61,
  UNRECOGNIZED = -1,
}
// ...
export const Index: MessageFns<Index> = {
  encode(message: Index, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.metadata !== undefined) {
      Metadata.encode(message.metadata, writer.uint32(10).fork()).join();
    }
    if (message.documents !== undefined && message.documents.length !== 0) {
      for (const v of message.documents) {
        Document.encode(v!, writer.uint32(18).fork()).join();
      }
    }
    if (message.externalSymbols !== undefined && message.externalSymbols.length !== 0) {
      for (const v of message.externalSymbols) {
        SymbolInformation.encode(v!, writer.uint32(26).fork()).join();
      }
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Index {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIndex();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.metadata = Metadata.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          const el = Document.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.documents!.push(el);
          }
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          const el = SymbolInformation.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.externalSymbols!.push(el);
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Index {
    return {
      metadata: isSet(object.metadata) ? Metadata.fromJSON(object.metadata) : undefined,
      documents: globalThis.Array.isArray(object?.documents)
        ? object.documents.map((e: any) => Document.fromJSON(e))
        : [],
      externalSymbols: globalThis.Array.isArray(object?.externalSymbols)
        ? object.externalSymbols.map((e: any) => SymbolInformation.fromJSON(e))
        : [],
    };
  },

  toJSON(message: Index): unknown {
    const obj: any = {};
    if (message.metadata !== undefined) {
      obj.metadata = Metadata.toJSON(message.metadata);
    }
    if (message.documents?.length) {
      obj.documents = message.documents.map((e) => Document.toJSON(e));
    }
    if (message.externalSymbols?.length) {
      obj.externalSymbols = message.externalSymbols.map((e) => SymbolInformation.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Index>, I>>(base?: I): Index {
    return Index.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Index>, I>>(object: I): Index {
    const message = createBaseIndex();
    message.metadata = (object.metadata !== undefined && object.metadata !== null)
      ? Metadata.fromPartial(object.metadata)
      : undefined;
    message.documents = object.documents?.map((e) => Document.fromPartial(e)) || [];
    message.externalSymbols = object.externalSymbols?.map((e) => SymbolInformation.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const SymbolInformation: MessageFns<SymbolInformation> = {
// ...
  fromJSON(object: any): SymbolInformation {
    return {
      symbol: isSet(object.symbol) ? globalThis.String(object.symbol) : "",
      documentation: globalThis.Array.isArray(object?.documentation)
        ? object.documentation.map((e: any) => globalThis.String(e))
        : [],
      relationships: globalThis.Array.isArray(object?.relationships)
        ? object.relationships.map((e: any) => Relationship.fromJSON(e))
        : [],
      kind: isSet(object.kind) ? symbolInformation_KindFromJSON(object.kind) : 0,
      displayName: isSet(object.displayName) ? globalThis.String(object.displayName) : "",
      signatureDocumentation: isSet(object.signatureDocumentation)
        ? Document.fromJSON(object.signatureDocumentation)
        : undefined,
      enclosingSymbol: isSet(object.enclosingSymbol) ? globalThis.String(object.enclosingSymbol) : "",
    };
  },
// ...
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}